pyproject.toml
[project]
name = "dbskill"
version = "0.1.0"
description = "Database skill toolkit with dual API/direct modes, permissions, and auditing."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "dbskill", email = "dev@example.com" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "pymysql>=1.1.0",
  "aiosqlite>=0.20.0",
  "PyYAML>=6.0.0",
  "passlib>=1.7.4",
  "python-multipart>=0.0.6",
  "itsdangerous>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["api*", "dbskill*"]
exclude = ["logs*"]
